添加面度识别功能：车型识别、动物识别、植物识别、菜品识别。

+++ app/src/main/java/com/example/jakera/smartchat/Entry/Car.java
@@ -7,7 +7,7 @@ import java.util.List;
  */
 
 public class Car {
-    public double log_id;
+    public long log_id;
     public List<CarInfo> result;
     public String color_result;
